ai_support_streamlit/requirements.txt
streamlit>=1.36.0
google-genai>=1.0.0
python-dotenv>=1.0.1
